backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.35
aiosqlite==0.20.0
cryptography==43.0.3
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
